New BinaryType for bool and str
